change_face/requirements.txt
opencv-python 
opencv-python-headless 
dlib 
numpy 
ffmpeg-python
